Fixed all type conversion warnings
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@6482 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DlgMediaVis.cpp
@@ -311,7 +311,7 @@ void CMediaVisDlg::OnOK()
 
 	if ( Settings.MediaPlayer.VisCLSID != strCLSID ||
 		 Settings.MediaPlayer.VisPath != strPath ||
-		 Settings.MediaPlayer.VisSize != m_nSize - 1 )
+		 Settings.MediaPlayer.VisSize != (DWORD)( m_nSize - 1 ) )
 	{
 		Settings.MediaPlayer.VisCLSID	= strCLSID;
 		Settings.MediaPlayer.VisPath	= strPath;
